bugfix: 404 if no key matches, fixes #6789
I also separated the parsing of the hkp response from
FetchKeyInfo.

This way FetchKeyInfo has the EM specific code that has
sideeffects and the logic is in a class without sideeffects
and (almost) without state.

The only state we keep is the KeyInfo array that contains
all the information the server returns. This way we avoid
parsing the response multiple times.