refactor: let handlers check if they are applicable
Instead of testing the preconditions for each handler in the dispatcher
the dispatcher hands a request to one handler after the other until one of
them responds.

This is similar to the Chain of Responsibility patter but we iterate over the
'handler_chain' array instead of a linked list.

To change the order of handlers or add other handlers change the array in the
handler_chain function.