deal with network issues in nicknym source
This is a first step. In case the suspected nicknym server cannot be
reached we will now move on and try other sources.

It's robably not what we want in the long run. In order to know wether
no key exists or we just failed to connect to some servers a different
http response code would be nice if network errors occured.

This simplifies testing such scenarios in the unit test and makes the
remote tests skip on network failure.