use plain http.rb rather than celluloid i/o
Turns out http.rb does not support celluloid i/o andymore and we were not making
use of it anyway.

Here's what https://github.com/httprb/http/wiki/Thread-Safety says:
(NOTE: this gem previously supported Celluloid::IO, but that support was removed to add the current timeout backend. It may be added back in a future version)

So now we have a plain http.rb adapter and use that as the default.
This prevents actors from crashing in the adapter (as we don't have any).