Version 0.9.0
Allow queries by fingerprint (using keyservers to reply) and support key
lookup from other nickservers.

This is a major rewrite both in terms of our dependencies and our
architecture.  We moved the server and http_client from eventmachine to
Celluloid based libraries. This allows for a more concise syntax and
proper https handling.

In addition we now support ruby 2.3, use a Logger instead of stdout and handle
network errors properly