fix: #3 handle domains without A-record
If a domain only has an mx record but no A record it
will trigger a ConnectionError when attempting the
nicknym key lookup.

We need to detect and handle this in the http adapter
already because once the exception is handled by Celluloid
our actor will be terminated.

So now we allow for handing a rescue option to the
adapter with a string that is checked for inclusion in
the error message. If the string is found the exception
will be caught and the adapter returns nil.

We only make use of this when checking the availability
of nicknym so far. That should be the only http request
going out.