fix: no expiration date means not outdated
We were using Time.at(expirationdate) even if it was nil which
led to using the Time.at(0). Instead an unset expirationdate
is meant to not expire the key at all.

Our tests did not catch this because the assertions were in
blocks that did not get run at all. (at least in the HKP
integration test).